requirements.txt
PIL
django==1.3.1
South==0.7.3
django-annoying==0.7.6
httpagentparser
-e hg+http://bitbucket.org/eXtractor/django-vkontakte-iframe/#egg=django-vkontakte-iframe
django-mediagenerator==1.11
-e hg+http://bitbucket.org/eXtractor/django-mediagenerator-closure-webservice#egg=django-mediagenerator-closure-webservice
#cssmin
#-e hg+http://bitbucket.org/eXtractor/django-mediagenerator-cssmin#egg=django-mediagenerator-cssmin
cssutils
-e git+https://github.com/myfreeweb/cssprefixer.git#egg=cssprefixer
-e hg+http://bitbucket.org/eXtractor/django-mediagenerator-cssprefixer#egg=django-mediagenerator-cssprefixer

textile==2.1.5
django-markitup==1.0.0

-e hg+https://bitbucket.org/mjs7231/django-dbbackup#egg=django-dbbackup
simples3
poster